Use YAML marshalling for config.yaml

+++ global/timeout.go
@@ -9,10 +9,10 @@ var TimeoutChannel chan bool = make(chan bool)
 /**
  * Returns a timeout value between average - jitter and average + jitter.
  */
-func GenerateTimeout(average int, jitter int) int {
+func GenerateTimeout(average uint32, jitter uint32) uint32 {
 	if jitter == 0 {
 		return average
 	}
 
-	return average - jitter + rand.Intn(2 * jitter)
+	return average - jitter + uint32(rand.Intn(2 * int(jitter)))
 }
